New db access function removing apps of a certain user

+++ src/server/service/service.cpp
@@ -347,16 +347,12 @@ error_label:
     return false;
 }
 
-bool Service::processAppUninstall(MessageBuffer &buffer, MessageBuffer &send, uid_t uid)
-{
-    // deserialize request data
-    std::string appId;
+
+bool Service::AppUninstall(uid_t uid, const std::string &appId) {
     std::string pkgId;
     std::string smackLabel;
     bool appExists = true;
     bool removePkg = false;
-
-    Deserialization::Deserialize(buffer, appId);
 
     try {
         std::vector<std::string> oldPkgPrivileges, newPkgPrivileges;
@@ -370,7 +366,7 @@ bool Service::processAppUninstall(MessageBuffer &buffer, MessageBuffer &send, ui
         } else {
             if (!generateAppLabel(pkgId, smackLabel)) {
                 LogError("Cannot generate Smack label for package: " << pkgId);
-                goto error_label;
+                return false;
             }
 
             std::string uidstr = isGlobalUser(uid) ? CYNARA_ADMIN_WILDCARD
@@ -391,35 +387,40 @@ bool Service::processAppUninstall(MessageBuffer &buffer, MessageBuffer &send, ui
     } catch (const PrivilegeDb::Exception::InternalError &e) {
         m_privilegeDb.RollbackTransaction();
         LogError("Error while removing application info from database: " << e.DumpToString());
-        goto error_label;
+        return false;
     } catch (const CynaraException::Base &e) {
         m_privilegeDb.RollbackTransaction();
         LogError("Error while setting Cynara rules for application: " << e.DumpToString());
-        goto error_label;
+        return false;
     } catch (const std::bad_alloc &e) {
         m_privilegeDb.RollbackTransaction();
         LogError("Memory allocation while setting Cynara rules for application: " << e.what());
-        goto error_label;
+        return false;
     }
 
-    if (appExists) {
-
-        if (removePkg) {
-            LogDebug("Removing Smack rules for deleted pkgId " << pkgId);
-            if (!SmackRules::uninstallPackageRules(pkgId)) {
-                LogError("Error on uninstallation of package-specific smack rules");
-                goto error_label;
-            }
+    if (appExists && removePkg) {
+        LogDebug("Removing Smack rules for deleted pkgId " << pkgId);
+        if (!SmackRules::uninstallPackageRules(pkgId)) {
+            LogError("Error on uninstallation of package-specific smack rules");
+            return false;
         }
     }
-
-    // success
-    Serialization::Serialize(send, SECURITY_MANAGER_API_SUCCESS);
     return true;
+}
 
-error_label:
-    Serialization::Serialize(send, SECURITY_MANAGER_API_ERROR_SERVER_ERROR);
-    return false;
+bool Service::processAppUninstall(MessageBuffer &buffer, MessageBuffer &send, uid_t uid)
+{
+    bool ret = false;
+    std::string appId;
+
+    Deserialization::Deserialize(buffer, appId);
+
+    ret = this->AppUninstall(uid, appId);
+    if (ret == true)
+        Serialization::Serialize(send, SECURITY_MANAGER_API_SUCCESS);
+    else
+        Serialization::Serialize(send, SECURITY_MANAGER_API_ERROR_SERVER_ERROR);
+    return ret;
 }
 
 bool Service::processGetPkgId(MessageBuffer &buffer, MessageBuffer &send)
@@ -517,5 +518,21 @@ bool Service::processGetAppGroups(MessageBuffer &buffer, MessageBuffer &send, ui
     return true;
 }
 
+
+bool Service::removeAppsOfUser(uid_t uid)//todo make client-side api for this?
+{
+    std::vector<std::string> userApps;
+    m_privilegeDb.GetUserApps(uid, userApps);
+    for (auto &app: userApps) {
+        //TODO is it ok to just break here in case of a error?
+        //maybe it would be better to pass vector of apps to AppUninstall,
+        //so we could rollback transaction in case of an error.
+        //on the orher hand: what should we do, when user removal is completed,
+        //and removal of users application failed?
+        if (AppUninstall(uid, app) == false)
+            return false;
+    }
+    return true;
+}
 
 } // namespace SecurityManager
